vinumc: Install include headers into a separate dir
It's better to keep the installed headers inside directory named after
the project to avoid naming conflicts.
