+++ subprojects/vinumc/tests/testlib.c
@@ -1,6 +1,7 @@
-#include "extern_library.h"
 #include <stdlib.h>
 #include <string.h>
+
+#include <vinumc/extern_library.h>
 
 struct return_value return_arg(call_ctx ctx) {
 	return ctx_get_text(ctx);
